requirements.txt
backports-abc==0.5
backports.functools-lru-cache==1.5
certifi==2019.3.9
chardet==3.0.4
cloudpickle==1.1.1
cycler==0.10.0
dask==1.2.2
decorator==4.4.0
enum34==1.1.6
future==0.17.1
futures==3.2.0
idna==2.8
imageio==2.5.0
kiwisolver==1.1.0
matplotlib==2.2.4
networkx==2.2
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.0.0
plyfile==0.7
-e git+git@gitlab.vision.in.tum.de:khamuham/ss19_extendingpointnet.git@64d58e7e4c38179d1a489f5681961a4221b73698#egg=pointnet&subdirectory=pointnet.pytorch
pptk==0.1.0
protobuf==3.7.1
pygame==1.9.6
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.0.1
requests==2.21.0
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
subprocess32==3.5.3
toolz==0.9.0
torch==1.1.0
torchfile==0.1.0
torchvision==0.3.0
tornado==5.1.1
tqdm==4.31.1
urllib3==1.24.3
visdom==0.1.8.8
websocket-client==0.56.0
